Don't pull down stage 0 from the internet
The composed build will set `PREBUILT_DOTNET_TOOL_DIR` to a location
where the dotnet tool can be copied from.

Long term we should get rid of the copy in favor of just using one
from a common location, but some of the Stage0 logic in the CLI build
system will need to change, I think.